% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankfull_flow.R
\name{bankfull_flow}
\alias{bankfull_flow}
\title{Bankfull Flow Calculation}
\usage{
bankfull_flow(xs, S, wsl = NA, options)
}
\arguments{
\item{xs}{cross-section of object class xsection}

\item{S}{bed slope in cross-section for use in Manning's equation (m/m)}

\item{wsl}{(optional) water surface level to set as bankfull for flow calculation (m)}

\item{options}{object of class rcr_options with options and constants for hydraulic calculations}
}
\value{
\item{flow_stage}{data frame with flow and stage data, cms and m, respectively}
}
\description{
Calculates the bankfull flow for a given cross-section. Manning's equation is used to
solve for the flow at a given stage, slope and cross-section properties.
}
\details{
The wsl parameter sets the water surface level for the flow calculation. If not provided, the function
will take the lower of the two bank stations at the bankfull elevation and solve for flow at that stage.
Note that in this context, the default calculation will be for 'bankfull' flow, however, any water surface
level can be inputted to this function to determine the flow at that stage.

A composite Manning's n value is used in estimation of the roughness value for the cross-section, which is used in
Manning's equation.
}
\examples{

}
\seealso{
\code{\link{normal_depth}} to estimate the normal depth using Manning's equation

\code{\link{flow_area}} to calculate the flow area in cross-section for given segments
}
\keyword{Manning}
\keyword{bankfull}
\keyword{flow}
\keyword{stage}
