% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcr_classes.R
\docType{class}
\name{bc-class}
\alias{bc-class}
\alias{bc}
\title{Boundary Condition Object}
\description{
bc:
A reference class to create a boundary conditions at a particular location in the model.
Note that currently only subcritical mode with normal depth boundary condition at the downstream-most
cross-section in the model is supported.
}
\section{Fields}{

\describe{
\item{\code{riverstation}}{name of the cross-section where the boundary condition is applied (character type)}

\item{\code{station}}{station location where the boundary condition is applied along profile (numeric type)}

\item{\code{reach}}{name of the reach}

\item{\code{location}}{description of the boundary condition location as either "Downstream", "Upstream"}

\item{\code{bctype}}{the type of boundary condition applied as "Normal Depth" (only supported type in current version)}

\item{\code{bcvalue}}{the value supplied with the boundary condition, currently the slope for normal depth computation}

\item{\code{init_WSL}}{the initial WSL value passed to boundary condition, used in normal depth calculation}
}}


