% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_flow_profiles.R
\name{compute_flow_profiles}
\alias{compute_flow_profiles}
\title{Hydraulic Profile Calculation for Multiple Flows}
\usage{
compute_flow_profiles(geometry, flows, boundary_conditions,
  method = "subcritical", options)
}
\arguments{
\item{geometry}{geometry object of class geom}

\item{flows}{vector of flow values for computation (cms)}

\item{boundary_conditions}{boundary conditions for the most downstream cross-section of object bc}

\item{method}{method used in calculation as "subcritical", "supercritical", or "mixed" (currently only "subcritical" method is supported)}

\item{options}{object of class rcr_options with options and constants for hydraulic calculations}
}
\value{
\item{hydraulic_output}{Table of computed hydraulic parameters by xsection}
}
\description{
Wrapper for the compute_profile function to run with multiple flow values and return an organized data frame
}
\details{
**Currently assumes subcritical regime throughout the profile, will be updated in future versions.**
}
\examples{

}
\seealso{
\code{\link{compute_profile}} to calculate the water profile for a single flow value
}
\keyword{compute}
\keyword{flow}
\keyword{profiles}
