% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_profile.R
\name{compute_profile}
\alias{compute_profile}
\title{Hydraulic Profile Calculation}
\usage{
compute_profile(geometry, Q, boundary_conditions, method = "subcritical",
  options)
}
\arguments{
\item{geometry}{geometry object of class geom}

\item{Q}{single flow value for computation (cms)}

\item{boundary_conditions}{boundary conditions for the most downstream cross-section of object bc}

\item{method}{method used in calculation as "subcritical", "supercritical", or "mixed" (currently only "subcritical" method is supported)}

\item{options}{object of class rcr_options with options and constants for hydraulic calculations}
}
\value{
\item{hydraulic_output}{Table of computed hydraulic parameters by xsection}
}
\description{
Calculates the water profile using the standard step method.
}
\details{
**Currently assumes subcritical regime throughout the profile, will be updated in future versions.**
}
\examples{

}
\seealso{
\code{\link{compute_flow_profiles}} to calculate flow profiles for multiple flow values
}
\keyword{compute}
\keyword{profile}
