% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_record_date_from_records.R
\name{extract_record_dates_from_record}
\alias{extract_record_dates_from_record}
\title{Extract Date from Record}
\usage{
extract_record_dates_from_record(record_path, all = F)
}
\arguments{
\item{record_path}{a file path to a record XML file}

\item{record_paths}{a vector of file paths to a record XML file}

\item{mc.cores}{the number of cores to use (Linux and Mac only) in \code{mclapply}.
Defaults to available cores - 1.}

\item{...}{further arguments supplied to \code{mclapply}.}
}
\value{
The function returns a \code{tibble} data frame with the following variables:
\describe{
  \item{record_id}{The id of the record.}
  \item{record_date}{The date of the record.}
}
}
\description{
The function extract date from the Riksdagen Records.
}
