% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_area.R
\name{flow_area}
\alias{flow_area}
\title{Flow Area Calculation}
\usage{
flow_area(xs, wsl, stns = NA, options)
}
\arguments{
\item{xs}{an object of class xsection}

\item{wsl}{an elevation at which the water level exists in the section}

\item{stns}{station points along the horizontal cross-section to divide the flow area. If none is supplied, the left and right bank stations only will be used.}

\item{options}{object of class rcr_options with options and constants for hydraulic calculations}
}
\value{
\item{flow_areas}{Table of computed cross-sectional flow area, m2}
}
\description{
Calculates the cross-sectional flow area
}
\details{
Calculates the flow area for a set of specified areas within the cross-section.
}
\examples{

}
\seealso{
\code{\link{bankfull_estimator}} to estimate the flow through the cross section using Manning's equation
}
\keyword{area}
\keyword{compute}
\keyword{flow}
