% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcr_classes.R
\docType{class}
\name{geom-class}
\alias{geom-class}
\alias{geom}
\title{Geometry Object}
\description{
geom:
A reference class to create a geometry object from a collection of cross-sections of class xsection
}
\section{Fields}{

\describe{
\item{\code{geomname}}{name of the geometry object}

\item{\code{xsectionList}}{a list of xsection objects, which contains all of the cross-sections and relevant data}
}}


