% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riksdag_corpora_path.R
\name{get_riksdag_corpora_path}
\alias{get_riksdag_corpora_path}
\alias{set_riksdag_corpora_path}
\title{Set and get the path to the Riksdag Corpora}
\usage{
get_riksdag_corpora_path()

set_riksdag_corpora_path(riksdag_corpora_path)
}
\arguments{
\item{riksdag_corpora_path}{The path}
}
\description{
Set and get the path to the Riksdag Corpora
}
