% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcr.R
\docType{package}
\name{rcr}
\alias{rcr}
\alias{rcr-package}
\title{rcr (Package)}
\description{
River Computations in R
}
\details{
This package has functions to calculate a basic backwater profile for open channels
using the standard step method.

\strong{***This package is currently under active devleopment to get to v1.0 and is not
yet fully implemented***}
}
\seealso{
\href{https://github.com/rchlumsk}{rchlumsk Github page} for more R packages.
}
