% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcr_classes.R
\docType{class}
\name{rcr_options-class}
\alias{rcr_options-class}
\alias{rcr_options}
\title{rcr Options Object}
\description{
rcr_options:
A reference class to define the computational parameters and physical constants used in the rcr package.
The parameters are defined with defaults when an instance of the options object is generated, however,
these may be modified as needed.
}
\section{Fields}{

\describe{
\item{\code{g}}{gravitational constant (N/kg)}

\item{\code{dx}}{resolution for interpolating cross-section data, defined by the horizontal space between interpolated points (m)}

\item{\code{tolerance_cp}}{tolerance limit for water surface calculations in compute_profile function (m)}

\item{\code{iteration_limit_cp}}{iteration limit for calculations between cross-sections in compute_profile}

\item{\code{next_WSL_split_cp}}{coefficient in assigning the next iteration of water surface elevations from the difference in previous WSL to calculated WSL}

\item{\code{tolerance_nd}}{tolerance limit for normal depth calculation in comparing flow to Manning's equation terms (m3/s)}

\item{\code{iteration_limit_nd}}{iteration limit for normal depth calculations}

\item{\code{next_WSL_split_nd}}{coefficient in assigning next water surface depth in normal depth calculation based on ratio of Manning's equation terms}

\item{\code{silent_cp}}{level of output in compute_profile function; if true, minimal output is produced}

\item{\code{silent_nd}}{level of output in normal_depth function; if true, minimal output is produced}
}}


