% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcr_support.R
\name{rcr_support}
\alias{rcr_support}
\alias{energy_calc}
\alias{sf_calc}
\alias{sf_calc2}
\alias{conv_calc}
\alias{vhead_calc}
\alias{wetted_perimeter}
\alias{froude_calc}
\title{Support functions for hydraulic calculations in the rcr package}
\usage{
energy_calc(Z, y, v)

sf_calc(n, v, Rh)

sf_calc2(Q, K)

conv_calc(n, A, Rh)

vhead_calc(alpha, v, g)

wetted_perimeter(zz, ind, dx)

froude_calc(v, g, d)
}
\arguments{
\item{Z}{water surface elevation (m)}

\item{y}{water depth (m)}

\item{v}{water average velocity (m/s)}

\item{n}{Manning's n value (weighted for channel)}

\item{Rh}{hydraulic radius (m)}

\item{Q}{flow (m3/s)}

\item{K}{channel conveyance

Other parameters and various outputs as per function usage.}
}
\description{
Support functions used in the rcr package for calculating various hydraulic properties.
Includes calculations for conveyance, friction slope, energy calculation, and others.
}
\details{
Note that some functions may have multiple versions for calcualting a hydraulic parameter,
such as conveyance or friction slope. To view the function calculation, type the function name
into the R console with no brackets to view the function code (e.g., 'energy_calc').
}
\keyword{calculations}
\keyword{functions}
\keyword{hydraulic}
\keyword{properties}
\keyword{rcr}
