% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo_sections.R
\docType{data}
\name{topo_sections}
\alias{topo_sections}
\title{Topographic data for Cross-sections}
\format{A data frame with four variables for three cross-sections:
\describe{
  \item{xsection}{cross-section number, lower numbers are downstream}
  \item{xx}{horizontal station coordinate, metres}
  \item{zz}{elevation in masl}
  \item{nn}{Manning's n value at station}
}}
\source{
hypothetical cross-sections
}
\usage{
topo_sections
}
\description{
A dataset containing the cross-section data for three hypothetical cross-secitons
for use in test cases.
}
\keyword{datasets}
