% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcr_classes.R
\docType{class}
\name{xsection-class}
\alias{xsection-class}
\alias{xsection}
\title{Cross-Section Object}
\description{
xsection:
A reference class to create a cross-section with all relevant data and static hydraulic properties.
}
\section{Fields}{

\describe{
\item{\code{riverstation}}{name of the cross-section as character}

\item{\code{station}}{station location along profile as numeric value}

\item{\code{xx}}{horizontal coordinates of 1D section}

\item{\code{zz}}{vertical elevation coordinates of 1D section, corresponding to xx}

\item{\code{Manning}}{vector of Manning's n values corresponding to xx (currently not used in rcr)}

\item{\code{Manning_LOB}}{Manning's n value for the left overbank area}

\item{\code{Manning_Main}}{Manning's n value for the main channel area}

\item{\code{Manning_ROB}}{Manning's n value for the right overbank area}

\item{\code{lbs_xx}}{horizontal co-orindate of the left bank station}

\item{\code{rbs_xx}}{horizontal co-orindate of the right bank station}

\item{\code{ds_length_LOB}}{downstream length of the left channel overbank}

\item{\code{ds_length_Main}}{downstream length of the main channel (to the next cross-section)}

\item{\code{ds_length_ROB}}{downstream length of the right channel overbank}

\item{\code{contraction_coeff}}{contraction coefficient for use in energy loss calculation}

\item{\code{expansion_coeff}}{expansion coefficient for use in energy loss calculation}
}}


