% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_speeches.R
\name{extract_speeches_from_record}
\alias{extract_speeches_from_record}
\title{Extract Speeches from Records}
\usage{
extract_speeches_from_record(record_path)
}
\arguments{
\item{record_path}{a file path to a record XML file}

\item{record_paths}{a vector of file paths to record XML files}
}
\value{
The function returns a tibble data frame with the following variables:
\describe{
  \item{speech_no}{The speech number in the record.}
  \item{speech_id}{The id of the XML node to the introduction of the speaker.}
  \item{who}{The id of the person giving the speech.}
  \item{id}{The id of the XML node for the segment of the speech.}
  \item{text}{The speech segment as plain text.}
}
}
\description{
The function extract speeches from the Riksdagen Records based on the
definition of a speech as the utterances (<u>) coming after a speaker
introduction (<note type="speaker">). The function returns the segments
of the speech.

For multiple files, parallelism can be used.
}
